rename toml module to toml_try_read
