+++ src/lib.rs
@@ -6,7 +6,7 @@ mod core {
     }
 
     mod config {
-        pub mod toml;
+        pub mod toml_try_read;
     }
 
     mod backup_execution;
